+++ src/main/java/com/easymart/request/CreateProductRequest.java
@@ -2,14 +2,15 @@ package com.easymart.request;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 @Data
 public class CreateProductRequest {
     private String title;
     private String description;
     private String brand;
-    private int mrpPrice;
-    private int sellingPrice;
+    private BigDecimal mrpPrice;
+    private BigDecimal sellingPrice;
     private String color;
     private List<String> images;
     private String category;
